list projects working
