--- server/.idea/dataSources/a317b321-eb7a-4794-a5ff-72e0b7e9cd7e/storage_v2/_src_/database/postgres.edMnLQ.meta
@@ -1,1 +1,0 @@
-#n:postgres
