--- server/.idea/dataSources/a317b321-eb7a-4794-a5ff-72e0b7e9cd7e/storage_v2/_src_/database/postgres.edMnLQ/schema/information_schema.FNRwLQ.meta
@@ -1,2 +1,0 @@
-#n:information_schema
-!<md> [null, 0, null, null, -2147483648, -2147483648]
