--- server/.idea/dataSources/a317b321-eb7a-4794-a5ff-72e0b7e9cd7e/storage_v2/_src_/database/postgres.edMnLQ/schema/pg_catalog.0S1ZNQ.meta
@@ -1,2 +1,0 @@
-#n:pg_catalog
-!<md> [null, 0, null, null, -2147483648, -2147483648]
