--- server/.idea/dataSources/a317b321-eb7a-4794-a5ff-72e0b7e9cd7e/storage_v2/_src_/database/postgres.edMnLQ/schema/public.abK9xQ.meta
@@ -1,2 +1,0 @@
-#n:public
-!<md> [494, 0, null, null, -2147483648, -2147483648]
