+++ server/Cargo.toml
@@ -1,15 +1,21 @@
+# CTRL+ALT+L
 [package]
 name = "server"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-serde = { version = "1.0", features = ["derive"] }
-anyhow = "1.0.51"
-actix-web = { version = "4.0.0-beta.13", features = ["rustls"] }
-actix-cors = "0.6.0-beta.6"
-sqlx = { version = "0.5", features = ["runtime-actix-rustls", "postgres", "uuid", "migrate", "macros", "offline", "chrono"] }
-uuid = "0.8.2"
+serde = { version = "1.0.132", features = ["derive"] }
+anyhow = "1.0.52"
+sqlx = { version = "0.5.9", features = ["runtime-tokio-rustls", "postgres", "uuid", "migrate", "macros", "offline", "chrono"] }
+uuid = { version = "0.8.2", features = ["v4"] }
 chrono = "0.4.19"
+axum = { version = "0.4.3", features = ["http2"] }
+hyper = { version = "0.14.16", features = ["full"] }
+tokio = { version = "1.15.0", features = ["full"] }
+tower = { version = "0.4.11" }
+tower-http = { version = "0.2.0", features = ["trace", "add-extension"] }
+tracing = "0.1.29"
+tracing-subscriber = { version = "0.3.4", features = ["env-filter"] }
+env_logger = "0.9.0"
+log = "0.4.14"
