+++ server/src/main.rs
@@ -1,33 +1,30 @@
-#![allow(dead_code)]
+use server::config::{Http, Postgres};
+use server::{config::Configuration, App};
 
-use anyhow::Result;
-use sqlx::postgres::{PgConnectOptions, PgPoolOptions, PgSslMode};
-use std::net::TcpListener;
+#[tokio::main]
+async fn main() -> hyper::Result<()> {
+    if std::env::var_os("RUST_LOG").is_none() {
+        std::env::set_var("RUST_LOG", "server=debug,tower_http=debug");
+    }
 
-#[actix_web::main]
-async fn main() -> Result<()> {
-    let pg_options = PgConnectOptions::new()
-        .username("postgres")
-        .password("postgres")
-        .host("127.0.0.1")
-        .port(5432)
-        .database("postgres")
-        .ssl_mode(PgSslMode::Prefer);
+    tracing_subscriber::fmt::init();
 
-    let _pg_pool = PgPoolOptions::new()
-        .connect_timeout(std::time::Duration::from_secs(2))
-        .connect_with(pg_options)
-        .await
-        .unwrap();
+    let config = Configuration {
+        http: Http {
+            host: "127.0.0.1".to_string(),
+            port: 8080,
+        },
+        postgres: Postgres {
+            host: "127.0.0.1".to_string(),
+            port: 5432,
+            user: "postgres".to_string(),
+            password: "postgres".to_string(),
+            database: "postgres".to_string(),
+            sslmode: false,
+        },
+    };
 
-    // TODO: move to App
-    // let config = Configuration::new()?;
-    let listener = TcpListener::bind("127.0.0.1:8080")?;
+    let app = App::new(config).await;
 
-    println!("Starting server... {}", "my string");
-
-    // TODO: let app = App::new(config) -> app.run()
-    server::run(listener)?.await?;
-
-    Ok(())
+    app.run().await
 }
