--- server/src/postgres.rs
@@ -1,1 +1,0 @@
-// https://github.com/vascokk/fullstack-rust/blob/main/server/src/main.rs
