--- server2/.env
@@ -1,6 +1,0 @@
-DATABASE_URL=postgres://postgres:postgres@localhost:5432/postgres
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=postgres
-MEILI_MASTER_KEY=key
-#SQLX_OFFLINE=true
-#POSTGRES_DB=my_databse
