--- server2/.idea/inspectionProfiles/Project_Default.xml
@@ -1,6 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="RsUnusedImport" enabled="true" level="WARNING" enabled_by_default="true" />
-  </profile>
-</component>
