--- server2/docker-compose.yaml
@@ -1,16 +1,0 @@
-version: '3.8'
-services:
-    postgres:
-      image: postgres:13.5
-      restart: always
-      env_file:
-        - .env
-      ports:
-        - '5432:5432'
-#    search:
-#      image: getmeili/meilisearch:latest
-#      restart: always
-#      env_file:
-#        - .env
-#      ports:
-#        - '7700:7700'
